Factor out common extrinsic submission
