+++ Cargo.toml
@@ -24,6 +24,7 @@ zip = { version = "0.5", default-features = false }
 pwasm-utils = "0.12"
 parity-wasm = "0.41"
 cargo_metadata = "0.9"
+codec = { package = "parity-scale-codec", version = "1.1" }
 
 # dependencies for optional extrinsics feature
 async-std = { version = "1.2.0", optional = true }
